Reader sidebar: switch from lib/user-settings to redux
