+++ client/reader/sidebar/index.jsx
@@ -6,12 +6,11 @@ import closest from 'component-closest';
 import createReactClass from 'create-react-class';
 import Gridicon from 'gridicons';
 import { localize } from 'i18n-calypso';
-import { defer, startsWith, identity, every } from 'lodash';
+import { defer, startsWith, identity } from 'lodash';
 import page from 'page';
 import React from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import store from 'store';
 
 /**
  * Internal dependencies
@@ -20,7 +19,6 @@ import ReaderSidebarHelper from './helper';
 import ReaderSidebarLists from './reader-sidebar-lists';
 import ReaderSidebarTags from './reader-sidebar-tags';
 import ReaderSidebarTeams from './reader-sidebar-teams';
-import AppPromo from 'blocks/app-promo';
 import QueryReaderLists from 'components/data/query-reader-lists';
 import QueryReaderTeams from 'components/data/query-reader-teams';
 import config from 'config';
@@ -29,10 +27,6 @@ import SidebarFooter from 'layout/sidebar/footer';
 import SidebarHeading from 'layout/sidebar/heading';
 import SidebarMenu from 'layout/sidebar/menu';
 import SidebarRegion from 'layout/sidebar/region';
-import observe from 'lib/mixins/data-observe';
-import userSettings from 'lib/user-settings';
-import userUtils from 'lib/user/utils';
-import { isMobile } from 'lib/viewport';
 import { isDiscoverEnabled } from 'reader/discover/helper';
 import { isAutomatticTeamMember } from 'reader/lib/teams';
 import { getTagStreamUrl } from 'reader/route';
@@ -41,10 +35,10 @@ import { getSubscribedLists } from 'state/reader/lists/selectors';
 import getReaderTeams from 'state/selectors/get-reader-teams';
 import { setNextLayoutFocus } from 'state/ui/layout-focus/actions';
 import { toggleReaderSidebarLists, toggleReaderSidebarTags } from 'state/ui/reader/sidebar/actions';
+import ReaderSidebarPromo from './promo';
 
 export const ReaderSidebar = createReactClass( {
 	displayName: 'ReaderSidebar',
-	mixins: [ observe( 'userSettings' ) ],
 
 	getInitialState() {
 		return {};
@@ -281,11 +275,7 @@ export const ReaderSidebar = createReactClass( {
 					/>
 				</SidebarRegion>
 
-				{ this.props.shouldRenderAppPromo && (
-					<div className="sidebar__app-promo">
-						<AppPromo location="reader" locale={ userUtils.getLocaleSlug() } />
-					</div>
-				) }
+				<ReaderSidebarPromo />
 
 				<SidebarFooter />
 			</Sidebar>
@@ -298,41 +288,12 @@ ReaderSidebar.defaultProps = {
 	translate: identity,
 };
 
-export const shouldRenderAppPromo = ( options = {} ) => {
-	// Until the user settings have loaded we'll indicate the user is is a
-	// desktop app user because until the user settings have loaded
-	// userSettings.getSetting( 'is_desktop_app_user' ) will return false which
-	// makes the app think the user isn't a desktop app user for a few seconds
-	// resulting in the AppPromo potentially flashing in then out as soon as
-	// the user settings does properly indicate that the user is one.
-	const haveUserSettingsLoaded = userSettings.getSetting( ' is_desktop_app_user' ) === null;
-	const {
-		isDesktopPromoDisabled = store.get( 'desktop_promo_disabled' ),
-		isViewportMobile = isMobile(),
-		isUserLocaleEnglish = 'en' === userUtils.getLocaleSlug(),
-		isDesktopPromoConfiguredToRun = config.isEnabled( 'desktop-promo' ),
-		isUserDesktopAppUser = haveUserSettingsLoaded ||
-			userSettings.getSetting( 'is_desktop_app_user' ),
-		isUserOnChromeOs = /\bCrOS\b/.test( navigator.userAgent ),
-	} = options;
-
-	return every( [
-		! isDesktopPromoDisabled,
-		isUserLocaleEnglish,
-		! isViewportMobile,
-		! isUserOnChromeOs,
-		isDesktopPromoConfiguredToRun,
-		! isUserDesktopAppUser,
-	] );
-};
-
 export default connect(
 	state => {
 		return {
 			isListsOpen: state.ui.reader.sidebar.isListsOpen,
 			isTagsOpen: state.ui.reader.sidebar.isTagsOpen,
 			subscribedLists: getSubscribedLists( state ),
-			shouldRenderAppPromo: shouldRenderAppPromo(),
 			teams: getReaderTeams( state ),
 		};
 	},
